add hook for useAirdropInput
